消息句柄升级到64位，编译通过未联调

+++ App/ServerModule/DataCache/DataCache.cs
@@ -117,7 +117,7 @@ class DataCache
     {
         LogSys.Log(ServerLogType.INFO, "{0}", msg);
     }
-    private void OnNameHandleChanged(bool addOrUpdate, string name, int handle)
+    private void OnNameHandleChanged(bool addOrUpdate, string name, ulong handle)
     {
         try {
             m_Channel.OnUpdateNameHandle(addOrUpdate, name, handle);
@@ -125,7 +125,7 @@ class DataCache
             LogSys.Log(ServerLogType.ERROR, "Exception {0}\n{1}", ex.Message, ex.StackTrace);
         }
     }
-    private void OnCommand(int src, int dest, string command)
+    private void OnCommand(ulong src, ulong dest, string command)
     {
         const string c_QuitDataStore = "QuitDataStore";
         const string c_ReloadConfig = "ReloadConfig";
@@ -145,7 +145,7 @@ class DataCache
             LogSys.Log(ServerLogType.ERROR, "Exception {0}\n{1}", ex.Message, ex.StackTrace);
         }
     }
-    private void OnMessage(uint seq, int source_handle, int dest_handle,
+    private void OnMessage(uint seq, ulong source_handle, ulong dest_handle,
         IntPtr data, int len)
     {
         try {
@@ -157,7 +157,7 @@ class DataCache
         }
     }
 
-    private void OnMessageResultCallback(uint seq, int src, int dest, int result)
+    private void OnMessageResultCallback(uint seq, ulong src, ulong dest, int result)
     {
 
     }
